LayoutTests/security/decode-buffer-size-expected.txt

ALERT: successfully caught exception TypeError: Type error
ALERT: successfully caught exception TypeError: Type error
ALERT: decoded string from array with length 0x80000000 populated with UTF-8 encoding of U+2118
ALERT: successfully decoded many buffers
